requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
discord.py==2.2.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
nest-asyncio==1.5.6
openai==0.27.7
PyYAML==6.0
requests==2.30.0
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
